feat(website): emote ticket table customization
